+++ l10n_br_account_product/models/l10n_br_account_product.py
@@ -26,21 +26,6 @@ NFE_IND_IE_DEST = [
 NFE_IND_IE_DEST_DEFAULT = NFE_IND_IE_DEST[0][0]
 
 
-class ImportDeclarationLine(models.Model):
-    _name = 'l10n_br_account_product.import.declaration.line'
-
-    import_declaration_id = fields.Many2one(
-        'l10n_br_account_product.import.declaration', u'DI',
-        ondelete='cascade', select=True)
-    sequence = fields.Integer(u'Sequência', default=1, required=True)
-    name = fields.Char(u'Adição', size=3, required=True)
-    manufacturer_code = fields.Char(
-        u'Código do Fabricante', size=60, required=True)
-    amount_discount = fields.Float(u'Valor',
-                                   digits=dp.get_precision('Account'),
-                                   default=0.00)
-
-
 class L10nBrIcmsRelief(models.Model):
 
     _name = 'l10n_br_account_product.icms_relief'
